Fixed linux compilation

+++ barcode/include/barcode/ExporterMenu.hpp
@@ -40,7 +40,7 @@ private:
 		std::string reverseEngineerPath();
 		int numberOfFiles();
 		void clean();
-		void render(int index, Hydra::World::IWorld* world);
+		void render(size_t index, Hydra::World::IWorld* world);
 	private:
 		std::string _name;
 		std::vector<Node*> subfolders;
